test(ci): colocate release-matrix trigger policy next to ci.yml
Move the deterministic trigger assertions out of scripts/ci-scope.test.ts
into .github/workflows/ci.test.ts (paired with ci.yml). Keeps release/**
push, rejects v* tags, and retains workflow_dispatch.
