+++ scripts/ci-scope.test.ts
@@ -199,42 +199,3 @@ test('changedFilesBetween keeps both sides of a cross-component rename', () => {
     rmSync(repo, { recursive: true, force: true });
   }
 });
-
-/**
- * Policy for the expensive six-job cross-platform matrix in ci.yml:
- * release/** push + workflow_dispatch only. v* tags must not re-trigger it —
- * they point at the commit already gated on the release branch.
- */
-describe('ci.yml release-matrix trigger policy', () => {
-  const ciYml = readFileSync(
-    join(import.meta.dir, '..', '.github', 'workflows', 'ci.yml'),
-    'utf8',
-  );
-
-  /** Slice the top-level `on:` block (lines between `on:` and `jobs:`). */
-  function onBlock(source: string): string {
-    const start = source.search(/^on:\s*$/m);
-    const jobs = source.search(/^jobs:\s*$/m);
-    expect(start).toBeGreaterThanOrEqual(0);
-    expect(jobs).toBeGreaterThan(start);
-    return source.slice(start, jobs);
-  }
-
-  test('push still triggers on release/** branches', () => {
-    const on = onBlock(ciYml);
-    expect(on).toMatch(/push:\s*\n\s+branches:\s*\['release\/\*\*'\]/);
-  });
-
-  test('v* tags do not trigger the matrix', () => {
-    const on = onBlock(ciYml);
-    // No tag filter under push (or anywhere in the on: block).
-    expect(on).not.toMatch(/tags:\s*\[/);
-    expect(on).not.toContain("tags: ['v*']");
-    expect(on).not.toContain('tags: ["v*"]');
-  });
-
-  test('workflow_dispatch remains for manual runs', () => {
-    const on = onBlock(ciYml);
-    expect(on).toMatch(/^\s*workflow_dispatch:\s*$/m);
-  });
-});
